pmix: Fix dropped completion in dUCX component
For the "send" codepath, the "buffer" feld is not
used. "msg" field is used instead.
Without the fix a random value in "buffer" determines
whether or not a completion event will be invoked.
All collectives are based on reliable completion delivery
so this bug could cause sporadic hangs when UCX component
is used.

Signed-off-by: Artem Polyakov <artpol84@gmail.com>